refactor(commands): extract command poll + dispatch + ack into commands/drain.ts (#1317)
§11e step 1 of COMMAND_PLUGIN_SKILL_ARCHITECTURE.md v1.0 (merged in
reflectt-cloud as commit da6cdba3). Behavior-zero relocation of the
inline command runtime that lived in cloud.ts.

Moved verbatim from cloud.ts into src/commands/drain.ts:
- pollAndProcessCommands()  — heartbeat-tick poll, 10s active / 60s idle
- handleCommand(cmd)        — inline if/else dispatch by cmd.type
- handleContextSync(cmd)    — fetches local context, posts to cloud sync
- handleRunApprove(cmd)     — calls local approval-queue, acks
- COMMAND_POLL_ACTIVE_MS / COMMAND_POLL_IDLE_MS constants
- commandPollErrors / lastCommandPollAt counters
- PendingCommand interface

Same poll cadence. Same ack contract. Same inline dispatch (no registry
yet — that lands in step 2). No new verbs. No auth/intake churn.

cloud.ts changes:
- exports state, config, isIdle, cloudGet, cloudPost so drain.ts can
  import them. ES live-binding semantics preserve current behavior;
  cloud.ts still owns the lets and reassigns config at line 368.
- imports pollAndProcessCommands + the two MS constants from
  ./commands/drain.js. Two existing call sites (heartbeat tick line 436
  and immediate-tick line 573) unchanged.
- 197 net lines removed; module shrinks toward owning just the
  cloud-transport concern.

Scope locks per @kai (msg-1777492135899) and @link (msg-1777492193718):
- behavior-zero
- same poll cadence
- same ack contract
- same inline dispatch semantics
- no registry creep until step 2
- no auth/intake churn
- no new verbs

Gates:
- typecheck: clean (npx tsc --noEmit, exit 0)
- vitest: 242/242 files pass, 2608/2608 tests pass (1 pre-existing skip)
- build: clean (npm run build)

Step 2 follow-up (per kai: "After that, reassess before the empty
registry PR") will land separately — this is step 1 only.
